Fix bug and refactor how struct members are resolved
Let's define the "left symbol" to refer to `a` in the expression `a.b`. For resolving struct members, there are 3 things to consider:
1. Unresolved left symbol (eg, in `a.b`, the program did not declare `a`)
2. Resolved left symbol but the StructSym corresponding to its type
doesn't exist (eg, `Blah a; a.b`, where `Blah` isn't defined anywhere
3. Resolved left symbol and its Type's corresponding StructSym but the right symbol is not a member

We must ensure `a` isn't an illegal forward reference, allow chained
members. (eg, `a.b.c` or `a.b().c()`)

Previously, we kept a stack of structs to track whether the left
symbol's Type was a StructSym or not. PrimaryId would push to the stack
if it were (if it weren't or is an illegal forward reference, it
wouldn't be pushed on). When memberAccess is called, it checks the stack. If it's empty,
it does nothing (this imples the `a` wasn't properly resolved and a
error was already broadcasted, so we didn't broadcast another one). If
the stack were not empty, it pops the StructSym, tries to resolve the
right symbol in it, and if it resolves to a VarSym with a type that has
a corresponding StructSym, pushes onto the stack.

This has a bug. Resolving members of builtintypes would not be reported
as an error. Eg) `int a; a.b` since the `a` wasn't pushed onto the
stack. In addition, no exception handling was added in case no
StructSymbol corresponded to the type. The compiler would crash. methodAccess also didn't handle chained calls like `a.b().c()`.

With this commit, we stop using the stack of structs and use the SymbolTable's
parseTreeContext map instead. Each node is placed in there, so that
memberAccess and methodAccess can get the left ParseTree node's corresponding
symbol, check if it's Type correspond to an existing StructSym
symbol, resolve members, and update the map. This fixes all the previous
bugs

We also create two new `ErrorEvent`s: `UNRESOLVED_STRUCTDEF_TO_MATCH_TYPE` and `BUILTINTYPE_CANNOT_HAVE_MEMBER`