Added codegen for if statements
Handles "if", "if/else", and recursive ones. The algorithm goes as
follows:

Upon entering the "if" statement, create and place the block labels if
"then", "else" (if an else block exists), and "merge". Then associate
the 'if'/'else' nodes with the blocks. Also pass in the expr and the
merge label. Note that the expr is evaluated after enterStmtIf is
called. This means we can't create the conditional br in enterStmtIf (b/c if we did
the expr's code would be printed after the br)

The insertion point for the "then" label can only be set by Then's
enterBlock. The insertion point for the "merge" label can only be set by
either Else's exitBlock or Then's exitBlock. As a result, upon entering
a block, if it's a then-node, retrieve the association to create the br, get comparison expr, etc.

Upon exiting a block, if it's a then or a else, retrieve the association to create the br to merge
point and set the insertion point to the next block (either merge or
else, depending on whether 'else' node exists or not, which is known by
the association)